refactor: Carousel, MissionList 첫글자 대문자로 변경
