+++ frontend/src/components/Carousel/Carousel.styled.ts
