+++ frontend/src/components/MissionList/MissionItem.tsx
