+++ frontend/src/components/MissionList/MissionList.styled.ts
