+++ frontend/src/components/missionDetailButtonGroup/index.tsx
@@ -1,6 +1,6 @@
 import { Link } from 'react-router-dom';
-import { missionMocks } from '../missionList/missionMocks';
-import type { Mission } from '../missionList/missionMocks';
+import { missionMocks } from '../MissionList/missionMocks';
+import type { Mission } from '../MissionList/missionMocks';
 import * as S from './MissionDetailButtonGroup.styled';
 
 interface MissionDetailButtonGroupProps {
