+++ frontend/src/components/missionDetailHeader/index.tsx
@@ -1,6 +1,6 @@
 import * as S from './missionDetailHeader.styled';
-import { missionMocks } from '../missionList/missionMocks';
-import type { Mission } from '../missionList/missionMocks';
+import { missionMocks } from '../MissionList/missionMocks';
+import type { Mission } from '../MissionList/missionMocks';
 
 interface MissionDetailHeaderProps {
   id: number;
